:fire: refactor: Admin, User 코드 리팩토링
