+++ backend/src/main/java/io/f1/backend/domain/admin/app/handler/AdminLoginFailureHandler.java
@@ -28,7 +28,7 @@ public class AdminLoginFailureHandler implements AuthenticationFailureHandler {
             HttpServletResponse response,
             AuthenticationException exception)
             throws IOException {
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED); // 401
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
         response.setContentType("application/json;charset=UTF-8");
 
         AdminLoginFailResponse errorResponse =
