+++ backend/src/main/java/io/f1/backend/domain/game/app/RoomService.java
@@ -10,8 +10,8 @@ import static io.f1.backend.domain.game.mapper.RoomMapper.toRoomSetting;
 import static io.f1.backend.domain.game.mapper.RoomMapper.toRoomSettingResponse;
 import static io.f1.backend.domain.game.websocket.WebSocketUtils.getDestination;
 import static io.f1.backend.domain.quiz.mapper.QuizMapper.toGameStartResponse;
-import static io.f1.backend.global.util.SecurityUtils.getCurrentUserId;
-import static io.f1.backend.global.util.SecurityUtils.getCurrentUserNickname;
+import static io.f1.backend.global.security.util.SecurityUtils.getCurrentUserId;
+import static io.f1.backend.global.security.util.SecurityUtils.getCurrentUserNickname;
 
 import io.f1.backend.domain.game.dto.MessageType;
 import io.f1.backend.domain.game.dto.RoomEventType;
