+++ backend/src/main/java/io/f1/backend/domain/user/app/UserService.java
@@ -19,7 +19,7 @@ import io.f1.backend.global.exception.CustomException;
 import io.f1.backend.global.exception.errorcode.AuthErrorCode;
 import io.f1.backend.global.exception.errorcode.UserErrorCode;
 import io.f1.backend.global.util.RedisPublisher;
-import io.f1.backend.global.util.SecurityUtils;
+import io.f1.backend.global.security.util.SecurityUtils;
 
 import jakarta.servlet.http.HttpSession;
