+++ backend/src/main/java/io/f1/backend/domain/user/app/handler/CustomAuthenticationEntryPoint.java
@@ -30,7 +30,7 @@ public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint 
             AuthenticationException authException)
             throws IOException {
         response.setContentType("application/json;charset=UTF-8");
-        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED); // 401
+        response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
 
         Map<String, String> errorResponse = new HashMap<>();
         errorResponse.put("code", AuthErrorCode.UNAUTHORIZED.getCode());
