+++ backend/src/test/java/io/f1/backend/domain/stat/RedisStatBrowserTest.java
@@ -33,7 +33,7 @@ import io.f1.backend.domain.user.dto.SignupRequest;
 import io.f1.backend.domain.user.entity.User;
 import io.f1.backend.global.config.RedisTestContainerConfig;
 import io.f1.backend.global.template.BrowserTestTemplate;
-import io.f1.backend.global.util.SecurityUtils;
+import io.f1.backend.global.security.util.SecurityUtils;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
